@@ -34,48 +34,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.1",
-    "@angular/cdk": "~9.2.0",
-    "@angular/common": "~9.1.1",
-    "@angular/compiler": "~9.1.1",
-    "@angular/core": "~9.1.1",
-    "@angular/fire": "^6.0.0",
-    "@angular/forms": "~9.1.1",
-    "@angular/material": "^9.2.0",
-    "@angular/platform-browser": "~9.1.1",
-    "@angular/platform-browser-dynamic": "~9.1.1",
-    "@angular/router": "~9.1.1",
+    "@angular/animations": "~11.0.0",
+    "@angular/cdk": "~11.0.0",
+    "@angular/common": "~11.0.0",
+    "@angular/compiler": "~11.0.0",
+    "@angular/core": "~11.0.0",
+    "@angular/fire": "^6.0.5",
+    "@angular/forms": "~11.0.0",
+    "@angular/material": "^11.0.0",
+    "@angular/platform-browser": "~11.0.0",
+    "@angular/platform-browser-dynamic": "~11.0.0",
+    "@angular/router": "~11.0.0",
     "angular2-qrcode": "^2.0.3",
-    "chart.js": "^2.9.3",
-    "firebase": "^7.14.0",
-    "hammerjs": "^2.0.8",
-    "mat-table-exporter": "^9.0.0",
-    "ng2-charts": "^2.3.0",
-    "ngx-order-pipe": "^2.0.4",
+    "chart.js": "^2.9.4",
+    "firebase": "^8.0.2",
+    "mat-table-exporter": "^9.0.2",
+    "ng2-charts": "^2.4.2",
+    "ngx-order-pipe": "^2.1.1",
     "randomatic": "^3.1.1",
-    "rxjs": "~6.5.5",
-    "tslib": "^1.11.0",
-    "zone.js": "~0.10.3"
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.1",
-    "@angular/cli": "~9.1.1",
-    "@angular/compiler-cli": "~9.1.1",
-    "@angular/language-service": "~9.1.1",
-    "@types/node": "~13.11.1",
-    "@types/jasmine": "~3.5.10",
+    "@angular-devkit/build-angular": "~0.1100.1",
+    "@angular/cli": "~11.0.1",
+    "@angular/compiler-cli": "~11.0.0",
+    "@types/firebase": "^3.2.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
-    "codelyzer": "^5.2.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.1",
-    "karma": "~5.0.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.1.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.1",
-    "karma-jasmine": "~3.1.1",
-    "karma-jasmine-html-reporter": "^1.5.3",
-    "protractor": "~5.4.3",
-    "ts-node": "~8.8.2",
-    "tslint": "~6.1.1",
-    "typescript": "~3.8.3"
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.2"
   }
 }
