@@ -22,19 +22,22 @@ import { MaterialModule } from './material.module';
 
 // Components
 import { AppComponent } from './app.component';
-import { AdminPanelComponent } from './components/admin-panel/admin-panel.component';
-import { CardsBoardComponent } from './components/cards-board/cards-board.component';
-import { ChartComponent } from './components/chart/chart.component';
-import { TicketsComponent } from './components/tickets/tickets.component';
-import { TicketsListComponent } from './components/tickets/tickets-list/tickets-list.component';
-import { HomeComponent } from './components/home/home.component';
-import { ShareComponent } from './components/share/share.component';
-import { ReportComponent } from './components/report/report.component';
-import { PopupComponent } from './components/share/popup/popup.component';
-import { RoomComponent } from './components/room/room.component';
-import { TextfieldPopupComponent } from './components/tickets/textfield-popup/textfield-popup.component';
-import { TicketPreviewComponent } from './components/room/ticket-preview/ticket-preview.component';
-import { UserDetailsComponent } from './components/room/user-details/user-details.component';
+import {
+  AdminPanelComponent,
+  CardsBoardComponent,
+  ChartComponent,
+  HomeComponent,
+  PopupComponent,
+  ReportComponent,
+  RoomComponent,
+  ShareComponent,
+  TextfieldPopupComponent,
+  TicketPreviewComponent,
+  TicketsComponent,
+  TicketsListComponent,
+  UserDetailsComponent,
+  UsersListComponent,
+} from './components';
 
 import { environment } from '../environments/environment';
 
@@ -43,17 +46,18 @@ import { environment } from '../environments/environment';
     AppComponent,
     AdminPanelComponent,
     CardsBoardComponent,
+    ChartComponent,
     HomeComponent,
-    ShareComponent,
     PopupComponent,
-    TicketsComponent,
-    TicketsListComponent,
     ReportComponent,
     RoomComponent,
-    ChartComponent,
+    ShareComponent,
     TextfieldPopupComponent,
     TicketPreviewComponent,
+    TicketsComponent,
+    TicketsListComponent,
     UserDetailsComponent,
+    UsersListComponent,
   ],
   imports: [
     BrowserModule,
