@@ -2,9 +2,7 @@ import { Component, Input, ChangeDetectionStrategy, EventEmitter, Output, OnChan
 import { MatDialog } from '@angular/material/dialog';
 import { MatSlideToggleChange } from '@angular/material/slide-toggle';
 
-import * as firebase from 'firebase/app';
-
-import { Ticket, User, Session } from 'src/app/models';
+import { Ticket, User, Session, FirestoreTimestamp, FirestoreFieldValue, fromDate, serverTimestamp } from 'src/app/models';
 import { TextfieldPopupComponent } from './textfield-popup/textfield-popup.component';
 import { HashService, SessionService, UrlParserService } from 'src/app/services';
 
@@ -44,11 +42,11 @@ export class TicketsComponent implements OnChanges {
   }
 
   onAddTicket(): void {
-    let timestamp:firebase.firestore.FieldValue = null;
+    let timestamp: FirestoreFieldValue = null;
 
     if (this.insertBefore && this.tickets.length) {
-      const nextDate = (this.tickets[0].timestamp as firebase.firestore.Timestamp).toDate();
-      timestamp = firebase.firestore.Timestamp.fromDate(new Date(Number(nextDate) - 10));
+      const nextDate = (this.tickets[0].timestamp as FirestoreTimestamp).toDate();
+      timestamp = fromDate(new Date(Number(nextDate) - 10));
     }
 
     this.sendTicket(this.ticket, timestamp);
@@ -78,11 +76,11 @@ export class TicketsComponent implements OnChanges {
     this.revote.emit(ticket);
   }
 
-  private sendTicket(title: string, timestamp?: firebase.firestore.FieldValue) {
+  private sendTicket(title: string, timestamp?: FirestoreFieldValue) {
     const ticket: Ticket = {
       sessionId: this.sessionService.getSessionId(),
       ticketId: this.hashService.generateHash(32),
-      timestamp: timestamp || firebase.firestore.FieldValue.serverTimestamp(),
+      timestamp: timestamp || serverTimestamp(),
       title,
       voted: false
     };
