@@ -1,7 +1,9 @@
+import { FirestoreFieldValue } from './firebase.model';
+
 export interface Session {
   id: string;
   title?: string;
   cardsType?: string;
-  timestamp?: firebase.firestore.FieldValue;
+  timestamp?: FirestoreFieldValue;
   activeTicket?: string;
 }
