@@ -1,8 +1,10 @@
+import { FirestoreFieldValue } from './firebase.model';
+
 export interface Ticket {
   title?: string;
   href?: string;
   sessionId?: string;
-  timestamp?: firebase.firestore.FieldValue;
+  timestamp?: FirestoreFieldValue;
   ticketId?: string;
   skipped?: boolean;
   voted?: boolean;
